==> 10-Master-Thesis/notebooks/00-Testing-Things.ipynb <==
Notebook dedicado a probar cosas que quiera implementar

from datetime import date

date.today()

date = date.today()

a = 'jeje'

date = date.strftime("%m-%d-%Y")

with open('%s_tweets.csv' % keyword, "w", newline='', encoding="utf-8") as f: 

print('%s_%s_tweets.csv'%(a, date))

import os 

directorio = os.getcwd()
os.path.join(directorio, 'tweets')
