==> 10-Master-Thesis/notebooks/01-Get-Twitter-Data.ipynb <==
# TFM Máster Data Science UAH 2020-2021

### MIGUEL PÉREZ CARO

Este notebook tiene como objetivo el desarrollo de un código capaz de obtener el dato de twitter que es necesario para la realización del proyecto. Se pretende obtener toda interacción que se haya producido con las 4 aerolíneas más importantes del mercado estadounidense que son las siguientes:

* Delta Airlines
* American Airlines
* Southwest Airlines
* United Airlines

### Extracción de Tweets

Importar librerías necesarias.

from datetime import date
import tweepy
import csv

### Autenticación

El primer paso son las credenciales

consumer_key = "xxxxxxx"
consumer_secret = "xxxxxxx"
access_token = "xxxxxxx" 
access_token_secret = "xxxxxxx" 

Posteriormente se prepara la autenticación

auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret) 

auth = tweepy.AppAuthHandler(consumer_key, consumer_secret)

Se prepara el módulo api de twitter que permitirá la posterior extracción de dato. Se usarán los siguientes parámetros para la llamada:

* compression: usa gzip para la compresión de los resultados.
* wait_on_rate_limit: automáticamente espera cuando se alcanza el límite.
* wait_on_rate_limit_notify: imprime por pantalla una notificación de que ha alcanzado el límite y está esperando a que pase el tiempo.

api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True, compression=True)

### Establecer campo de búsqueda

A continuación se especifica la palabra a buscar, el número de tweets y el lenguaje. En cuanto a los términos a usar para la búsqueda de tweets se van a seguir dos métodos de búsqueda para las 4 aerolíneas:

* Por un lado, se va a buscar con el nombre de usuario de la aerolínea tal que @nombre-usuario
* La segunda opción será el nombre de la aerolíneas acompañada de la palabra airlines.

De esta forma, se cubre un gran espectro de las posibles formas de interactuar con la compañía, ya sea mencionándola, o por medio del nombre completo de la aerolínea. El uso exclusivo del nombre principal de la aerolínea no se considera viable ya que, por ejemplo, la aerolínea United, es un nombre muy genérico, o la aerolínea Delta, es una palabra muy usada ya que así se ha denominado a una de las variantes del virus covid-19 por lo que se descarta esta opción.

Los términos a buscar son:

keywords = ["@AmericanAir", "American Airlines", 
            "@Delta", "Delta Airlines", 
            "@united", "United Airlines", 
            "@SouthwestAir", "Southwest Airlines"]

# Set the max number of tweets
num_tweets = 15000

# Set the language
language = 'en'

# Set the date
today = date.today()
today = today.strftime("%m-%d-%Y")

### Extracción Final

Para la extracción final se genera una función que se encarga de recoger los tweets para la palabra de búsqueda correspondiente.

def get_tweets(keyword,language, num_tweets):
    """
    Function used to get the tweets.
    
    - keyword: parameter to set the keyword we are looking for.
    - language: the language of the tweets.
    - num_tweets: the number of tweets we are looking for.
    """
    
    # List to store the raw tweets
    tweets = []
    print("Descargando tweets")
    
    # Used of tweepy api to get the tweets
    for tweet in tweepy.Cursor(api.search, keyword, lang=language).items(num_tweets): 
        try: 
            tweets.append(tweet) 
            print(len(tweets))
        except tweepy.TweepError as e: 
            print(e.reason) 
        except StopIteration: 
            break 
    
    # List to store the information we need from the raw tweet
    output = [(tweet.id_str, tweet.created_at, tweet.text,
                  tweet.retweet_count,str(tweet.favorite_count)) for tweet in tweets] 

    # Write it to a csv file
    with open('%s_tweets_%s.csv' %(keyword, today) , "w", newline='', encoding="utf-8") as file: 
        writer = csv.writer(file, quoting=csv.QUOTE_ALL) 
        writer.writerow(['id','created_at','text','retweet_count','favorite_count']) 
        writer.writerows(output) 
    pass 
    return 'Se han extraído los Tweets correctamente' 

Una vez creada la función, se l¡genera un bucle que pasa por los diferentes términos a buscar y genera una archivo csv para cada uno de ellos.

for keyword in keywords[1:]:
    print(keyword)
    get_tweets(keyword,language,num_tweets)

### Referencias

* Tweepy library documentation: https://docs.tweepy.org/en/stable/api.html
* CSV library documentation: https://docs.python.org/3/library/csv.html
* API Twitter documentation: https://developer.twitter.com/en/docs/twitter-api
* Ord function documentation: https://www.programiz.com/python-programming/methods/built-in/ord
* Passenger US Airlines: https://es.statista.com/estadisticas/598766/las-aerolineas-lideres-en-los-ee-uu-cuota-de-mercado-nacional/
