==> 10-Master-Thesis/notebooks/02-Join-Files.ipynb <==
# TFM Máster Data Science UAH 2020-2021

### MIGUEL PÉREZ CARO

Este notebook tiene como objetivo juntar todos los archivos individuales generados para la recogida de tweets y eliminar los duplicados qu epueda existir, de forma que se obtenga un set de datos final, sin duplicados y sin preprocesar.

Importar librerías

import os
import glob
import pandas as pd

Establecer drectorio de trabajo

directorio = os.path.dirname(os.getcwd())
mydir = os.path.join(directorio, 'data')
os.chdir(mydir)

Obtener todos los archivos a partir de la extensión csv:

extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]

Concatenar los archivos en uno solo

#df_raw_dup = pd.concat([pd.read_csv(f) for f in all_filenames])

def concatenate(filenames):
    df_raw_dup = pd.DataFrame()
    for f in filenames:
        if 'American' in f:
            df = pd.read_csv(f)
            df['american'] = 1
            df_raw_dup = pd.concat([df_raw_dup, df])
        
        if 'Delta' in f:
            df = pd.read_csv(f)
            df['delta'] = 1
            df_raw_dup = pd.concat([df_raw_dup, df])
        
        if 'Southwest' in f:
            df = pd.read_csv(f)
            df['southwest'] = 1
            df_raw_dup = pd.concat([df_raw_dup, df])
        
        if 'United' in f or 'united' in f:
            df = pd.read_csv(f)
            df['united'] = 1
            df_raw_dup = pd.concat([df_raw_dup, df])
    
    return df_raw_dup

df_raw_dup = concatenate(all_filenames)

Comprobar el tamaño del dataset

df_raw_dup.shape

df_raw_dup.info()

Comprobar los tweets duplicados por id

print(df_raw_dup['id'].duplicated().sum())

Estos duplicados se deben a la metodología de extracción de datos, con 3 posibles escenarios:

- Para la recogida de tweets de una aerolínea, se usan dos formatos, por el username, y por el nombre de la aerolínea. Es posible que un mismo tweet mencione la aerlínea por su username y por su nombre, lo que haría que existiese un duplicado.
- Existe la posibilidad de que un mismo tweet mencione 2,3 o las 4 aerolíneas, por lo que sería recogido en cada set de datos correspondiente a cada aerolínea.
- Finalmente, también puede suceder, debido a los tramos de tiempo de recogida, que un mismo tweet aparezca en dos set de datos que se hayan recogido en diferentes fechas.

En el primer y en el último escenario, la eliminación de todos los duplicados menos uno sería suficiente para arreglar el problema, pero en el segundo no, ya que aparecería solo 1 tweet correspondiente a una de las aerolíneas, pero no tendríamos constancia del resto, por lo que hay que tenerlo en cuenta a la hora de tratar el dato. Para ello, en vez de una elimnación de duplicados, se va a realizar un groupby por el id.

En primer lugar observamos el dataset:

df_raw_dup.head()

Rellenar los valores nulos con un 0

df_raw_dup['american'] = df_raw_dup['american'].fillna(0)
df_raw_dup['delta'] = df_raw_dup['delta'].fillna(0)
df_raw_dup['southwest'] = df_raw_dup['southwest'].fillna(0)
df_raw_dup['united'] = df_raw_dup['united'].fillna(0)

Se realiza la operación de groupby

df = df_raw_dup.groupby('id').agg({'created_at':'first', 
                           'text': 'first', 
                           'retweet_count':'max',
                           'favorite_count': 'max',
                           'southwest': 'max',
                           'american': 'max',
                           'united': 'max',
                           'delta': 'max'}).reset_index()

Se comprueba el resultado:

df.shape

df.head()

print(df['id'].duplicated().sum())

Parece que ya se ha obtenido un dataset sin duplicados, por lo que se procede a su almacenamiento.

Cabe destacar que el objetivo de este notebook es únicamente el de juntar los set de datos y guardarlos en un único dataset, por lo que aún queda realizar una parte del preprocesado del dato, que incluso incluirá tratamiento de duplicados, ya que en este caso solo se ha tenido en cuenta el tweet id, es decir, el identificador de cada tweet, pero puede haber muchos tweets cuyo texto sea igual.

Guardar el dataset en sin preprocesar

df.to_csv("df_raw.csv", sep = ',', index=False, 
              columns = ['id','created_at','text','retweet_count','favorite_count', 'southwest', 
                         'american', 'united', 'delta'])

### Referencias

* Glob library documentation: https://docs.python.org/3/library/glob.html
* OS library documentation: https://docs.python.org/3/library/os.path.html
* Post: https://www.freecodecamp.org/news/how-to-combine-multiple-csv-files-with-8-lines-of-code-265183e0854/
