==> 10-Master-Thesis/notebooks/04-Clean-Data.ipynb <==
# TFM Máster Data Science UAH 2020-2021

### MIGUEL PÉREZ CARO

Este notebook tiene como objetivo limpiar el dataset para que pueda ser usado posteriormente. Tratamiento del texto:

- Convertir a minúsculas
- Eliminar stopwords
- Eliminar signos de puntuación y símbolos extraños
- Eliminar links
- Tratar los espacios
- Eliminar otro tipo de palabras frecuentes que no aporten valor.

En primer lugar se importan las librerías.

import os
import re
import string
import numpy as np
import pandas as pd
import seaborn as sns
from nltk.corpus import stopwords
from emot.emo_unicode import EMOTICONS_EMO, UNICODE_EMOJI

import warnings
warnings.filterwarnings('ignore')

Se establece el directorio de trabajo.

directorio = os.path.dirname(os.getcwd())
mydir = os.path.join(directorio, 'data')
os.chdir(mydir)

Se carga el archivo

df_raw = pd.read_csv('df_raw.csv')

Se modifica el tipo de los datos

df_raw['id'] = pd.to_numeric(df_raw['id'], errors = 'coerce')
df_raw.dropna(inplace=True)
df_raw['created_at'] = df_raw['created_at'].astype('datetime64[ns]')
df_raw['retweet_count'] = df_raw['retweet_count'].astype('float64')

df_raw.info()

Se analizan los textos duplicados:

df_raw['text'].duplicated().sum()

Se eliminan los textos duplicados en base a las conclusiones obtenidas en el notebook 03-Analyze-Data. Para ello se hace uso de la función groupby

df_raw_no_dup = df_raw.groupby('text').agg({'id': 'first', 
                                            'created_at':'first', 
                                            'retweet_count':'max',
                                            'favorite_count': 'max',
                                            'southwest': 'max',
                                            'american': 'max',
                                            'united': 'max',
                                            'delta': 'max'}).reset_index()

Se comprueba el resultado

df_raw_no_dup['text'].duplicated().sum()

df_raw_no_dup.shape

Se convierte el id a un string

df_raw_no_dup['id'] = df_raw_no_dup['id'].astype('object')

df_raw_no_dup.info()

df_raw_no_dup.shape

Se genera un csv sin limpiar pero sin texto duplicado.

df_raw_no_dup.to_csv("df_raw_no_dup.csv", sep = ',', index=False, 
              columns = ['id','created_at','text','retweet_count','favorite_count', 
                         'southwest', 'american', 'united', 'delta'])

Para comenzar con la limpieza, se van a generar en primer lugar dos funciones para la limpieza de los emojis y los emoticonos, para continuar con otra función que incluir'a una limpieza más general y llamar'a a estas dos funciones.

def clean_emoticons(text):
    """
    Función generada para hacer la limpieza de emoticonos.
    
    param: text texto a limpiar
    :return: texto limpio
    """
    for emot in EMOTICONS_EMO:
        if emot in text:
            text = text.replace(emot, " {} ".format(EMOTICONS_EMO[emot]))
    return text

def clean_emojis(text):
    """
    Función generada para hacer la limpieza de emojis.
    
    param: text texto a limpiar
    :return: texto limpio
    """
    for emot in UNICODE_EMOJI:
        if emot in text:
            text = text.replace(emot, " {} ".format(UNICODE_EMOJI[emot].replace(':',"").replace('_',' ')))
    return text

Se procede ahora con la limpieza para el análisis de sentimiento. Para ello, se crea una función que realiza lo siguiente:

- Eliminar los usuarios. No aportan información al  análisis de sentimiento que se pretende hacer posteriormente.
- Eliminar los enlaces web.
- Convertir todo el texto a minúsculas.
- Eliminar los emojis y emoticonos. Importante que se eliminen tras la limpieza de los enlaces web ya que estos pueden incluir secuencias de carácteres idénticas a algún emoticono, lo que conllevaría que se preprocesaría como si fuera un emoticono aunque fuera parte de un enlace web.
- Eliminar los signos de almohadilla, ya que el hashtag si que puede contener información. Eliminar la palabra rt, que no aporta valor y aparece con frecuencia. 
- Eliminar los nombres de las aerolíneas que se usaron para la búsqueda ya que tampoco aporta ningún valor al análisis de sentimiento y ya está recogida su presencia en el texto. En estos se incluyen también acrónimos, y las siglas de los números de vuelo. Por ejemplo un vuelo de American Airlines podría ser el AA123, que tras la limpieza resultaría en aa, que no aporta ningún valor.
- Eliminar las abreviaturas usadas para situar a los aviones en vuelo, ya que tampoco aportan valor al texto. En estos casos se incluyen términos como mi, para describir las millas, ft, que son los pies de altura, o frm y hrzn para situar las coordenadas del avión.
- Eliminar lo que no sea minúscula o espacio.
- Normalizar espacios.
- Eliminar stopwords.

lista_elementos_eliminar = ['#', ' rt ', '\n', 'american airlines', 'delta airlines', 'united airlines', 
                            'southwest airlines', ' aa ', ' aal ', ' dl ', ' dal ', ' ua ', ' ual ', ' swa ',
                            ' mi ', ' ft ', ' frm ', ' hrzn ', ' amp ', ' aaa ']

def cleaning_text_basico(df, lista_elementos_eliminar):
    """
    Función generada para hacer una limpieza general del texto.
    
    param: df dataframe a limpiar
    :return: lista con el texto limpiado. Un elemento por cada fila del dataframe.
    """
    corpus = [' ' + text + ' ' for text in  df.text.tolist()] # Añadir espacio al principio y al final apra fcilitar limpieza
    corpus = [re.sub('@[A-Za-z0-9_]+',' ', text) for text in corpus] # Eliminar usuarios
    corpus = [re.sub(r'http\S+', '', text) for text in corpus] # Eliminas enlaces web
    corpus = [text.lower() for text in corpus] # Pasar a minúsculas
    #corpus = [clean_emojis(text) for text in corpus] # Eliminar emojis
    #corpus = [clean_emoticons(text) for text in corpus] # Eliminar emoticonos
    #corpus = [''.join(ch for ch in text if ch not in string.punctuation) for text in corpus] # Eliminar signos de puntuación
    corpus = [re.sub(r'[^a-z\s]', '', text) for text in corpus] # Eliminar lo que no sea minúscula o espacio
    for i in lista_elementos_eliminar:
        corpus = [re.sub(i,' ', text) for text in corpus] # Eliminar elementos de la lista
    corpus = [text.strip() for text in corpus] # Normalizar espacios
    corpus = [re.sub(r' +', ' ', text) for text in corpus] # Normalizar espacios
    corpus = [' '.join(word for word in text.split() if word not in stopwords.words('english')) for text in corpus] # Eliminar stopwords 
    return corpus

corpus_clean = cleaning_text_basico(df_raw_no_dup, lista_elementos_eliminar)

corpus_clean

Se genera un nuevo dataframe con el contenido limpio

df_clean = df_raw_no_dup.copy()
df_clean['text'] = corpus_clean

Debido a la limpieza, es posible que algunos textos que solo contuvieran un usuario, o el usuario y un enlace web, sean ahora textos vacíos, o solo espacios, por lo que es necesario tratarlo. En primer lugar, se convierten los textos que se compongan únicamente de espacios a valores nulos.

df_clean['text'] = df_clean['text'].replace(r'^\s*$', np.NaN, regex=True)

Se comprueban ahora los valores nulos y se eliminan.

df_clean['text'].isnull().sum()

df_clean.dropna(inplace=True)

Se comprueba si existen textos duplicados.

df_clean['text'].duplicated().sum()

Sí que hay textos duplicados. Esto se puede deber a múltiples factores como por ejemplo:

- Textos cuya única diferencia sea el RT al principio del mismo.
- Textos cuya única diferencia sea la mención a un usuario.
- Textos que contengan links a noticias de forma que el tweet es el mismo pero con un link diferente.

Todo ello puede hacer que tras la limpieza se encuentren nuevamente textos duplicados y que hay que tratar de la misma manera que anteriormente

df_clean_no_dup = df_clean.groupby('text').agg({'id': 'first', 
                                            'created_at':'first', 
                                            'retweet_count':'max',
                                            'favorite_count': 'max',
                                            'southwest': 'max',
                                            'american': 'max',
                                            'united': 'max',
                                            'delta': 'max'}).reset_index()

Se comprueba el nuevo dataframe generado

df_clean_no_dup['text'].duplicated().sum()

df_clean_no_dup.info()

df_clean_no_dup.shape

A continuación, se muestran por pantalla las palabras más y menos comunes del texto para detectar alguna anomalía que se pueda repetir tanto que sea necesario tratarla.

pd.Series(' '.join(df_clean_no_dup['text']).lower().split()).value_counts()[:50]

pd.Series(' '.join(df_clean['text']).lower().split()).value_counts()[-50:]

Analizando las palabras más comunes, no hay ningún término que llame la atención, lo cuál podría significar algún tipo de fallo a la hora de preprocesar el texto, por lo que es una buena señal el resultado obtenido. 

En cuanto a las palabras menos comunes, se puede observar que muchas de ellas son fallos a la hora de escribir los vocablos, por lo que podrían ser eliminados, pero no se considera necesario, y se decide no modificar el texto.

df_clean_no_dup.to_csv("df_sentiment_clean_no_dup.csv", sep = ',', index=False, 
              columns = ['id','created_at','text','retweet_count','favorite_count', 
                         'southwest', 'american', 'united', 'delta'])
