==> 10-Master-Thesis/notebooks/08-tfm-bert-model.ipynb <==
# TFM Máster Data Science UAH 2020-2021

### MIGUEL PÉREZ CARO

Este notebook tiene como objetivo llevar a cabo el entrenamiento del modelo BERT para convertir el texto a vector a partir de este modelo dado que no se puede realizar en local por falta de medios.

Se comienza por importar las librerías necesarias

!pip install transformers

import time
import torch
import numpy as np
import pandas as pd
import transformers as ppb

Se lleva a cabo la importación del dataset

df = pd.read_csv('df_clean_no_dup.csv')

df.drop(columns=['id', 'created_at', 'retweet_count', 'favorite_count'], inplace = True)

df = df[df['text'].apply(lambda x: len(x.split(' ')) > 4)]
df.reset_index(drop=True, inplace=True)

df.sample(5)

n = 400  #chunk row size
df=df[98400:]
list_df = [df[i:i+n] for i in range(0,df.shape[0],n)]

Se procede a trabajar con el modelo BERT

model_class, tokenizer_class, pretrained_weights = (ppb.BertModel, ppb.BertTokenizer, 'bert-base-uncased')

# Cargar model/tokenizer
tokenizer = tokenizer_class.from_pretrained(pretrained_weights)
model = model_class.from_pretrained(pretrained_weights)

bert_array = np.zeros((0,768))

for iteration, item in enumerate(list_df):
  tokenized = item['text'].apply((lambda x: tokenizer.encode(x, add_special_tokens=True)))
  # Obtener lista más larga
  max_len = len(max(tokenized, key=len))

  # Generar el array
  array_tokenized = np.array([i + [0]*(max_len-len(i)) for i in tokenized.values])

  # Generar el attention_mas
  attention_mask = np.where(array_tokenized != 0, 1, 0)

  array_tokenized_input = torch.tensor(array_tokenized)  
  attention_mask = torch.tensor(attention_mask)

  with torch.no_grad():
    bert_corpus_full = model(array_tokenized_input, attention_mask=attention_mask)
  
  bert_vectors = bert_corpus_full[0][:,0,:].numpy()

  bert_corpus = np.zeros((len(bert_vectors), 768))

  for idx, vector in enumerate(bert_vectors):
    bert_corpus[idx, :] = vector
  bert_array = np.concatenate((bert_array, bert_corpus), axis = 0)
  
  if iteration%10==0:
    print(iteration)


np.savetxt("bert_corpus_2.csv", bert_array, delimiter=",")


from google.colab import files
import pandas as pd
#result.to_csv('bert_corpus_2.csv')
files.download('bert_corpus_2.csv')
