==> 10-Master-Thesis/notebooks/09-Análisis-Dataset-Kaggle.ipynb <==
# TFM Máster Data Science UAH 2020-2021

### MIGUEL PÉREZ CARO

Este notebook tiene como objetivo el análisis de un set de datos obtenido a través de Twitter en Febrero de 2015 que recoge tweets que hacen referencia a las principales aerolíneas estadounidenses y cuyo sentimiento ha sido clasificado manualmente. El set de datos se encuentra en: https://www.kaggle.com/crowdflower/twitter-airline-sentiment

EL objetivo es comprobar dicho sentimiento y las diferencias que se encuentran con los resultados obtenidos en el dataset que se ha recogido durante el verano de 2021.

En primer lugar se importan las librerías

import os
import re
import string
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from emot.emo_unicode import EMOTICONS_EMO, UNICODE_EMOJI
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

Se establece el directorio de trabajo.

directorio = os.path.dirname(os.getcwd())
mydir = os.path.join(directorio, 'data')
os.chdir(mydir)

Se carga el archivo.

df = pd.read_csv('kaggle.csv')

Se analiza el dataset.

df.head()

df.shape

df.info()

df.airline.unique()

Solo interesan los campos:

- airline_sentiment
- negativereason
- airline
- text

Y las aerolíneas analizadas en el estudio que son: American, Delta, Southwest y United

df.drop(columns=['tweet_id', 'airline_sentiment_confidence', 'negativereason_confidence', 'airline_sentiment_gold',
                 'name', 'negativereason_gold', 'retweet_count', 'tweet_coord', 'tweet_created', 'tweet_location',
                 'user_timezone'], inplace = True)

df = df[df['airline'].isin(['United', 'Southwest', 'Delta', 'American'])]

df.airline.unique()

Se va a comprobar la distribución del sentimiento de los tweets por aerolínea:

print("Cantidad total de tweets recogidos por aerolínea:\n")
print(df.groupby('airline')['text'].count())

airlines= ['United','American','Southwest','Delta']
columns1 = ['Positivo', 'Neutral', 'Negativo', 'Total']
columns2 = ['Positivo', 'Neutral', 'Negativo']

data_total = list()
data_porcentaje = list()

for i in airlines :

    data1 = [list(df.loc[df.airline == i].airline_sentiment.value_counts().sort_index(ascending = False).values)+[df.loc[df.airline == i].shape[0]]]
    data_total = data_total+data1
    
    data2 = np.array(df.loc[df.airline == i].airline_sentiment.value_counts().sort_index(ascending = False).values)/df.loc[df.airline == i].shape[0]
    data2 = np.around(data2, decimals=4)*100
    data_porcentaje.append(data2)  

df_total = pd.DataFrame(data_total, columns = columns1, index = airlines)
df_total

df_porcentaje = pd.DataFrame(data_porcentaje, columns = columns2, index = airlines)
df_porcentaje

Una vez mostrados los resultados en cuanto al sentimiento por aerolínea, también interesa obtener los diferentes motivos a los que se debe el sentimiento negativo, lo cuál puede ser comparado con los patrones encontrados:

print('Razones por las que un tweet se ha considerado como algo negativo:\n')
print(df.negativereason.value_counts())

Por curiosidad, se va a aplicar la limpieza realizada y a utilizar el analizador de Sentimientos VADER para ver si los resultados difieren en gran medida de los obtenidos.

Se comienza por la limpieza

def clean_emoticons(text):
    """
    Función generada para hacer la limpieza de emoticonos.
    
    param: text texto a limpiar
    :return: texto limpio
    """
    for emot in EMOTICONS_EMO:
        if emot in text:
            text = text.replace(emot, " {} ".format(EMOTICONS_EMO[emot]))
    return text

def clean_emojis(text):
    """
    Función generada para hacer la limpieza de emojis.
    
    param: text texto a limpiar
    :return: texto limpio
    """
    for emot in UNICODE_EMOJI:
        if emot in text:
            text = text.replace(emot, " {} ".format(UNICODE_EMOJI[emot].replace(':',"").replace('_',' ')))
    return text

lista_elementos_eliminar = ['#', ' rt ', '\n', 'american airlines', 'delta airlines', 'united airlines', 
                            'southwest airlines', ' aa ', ' aal ', ' dl ', ' dal ', ' ua ', ' ual ', ' swa ',
                            ' mi ', ' ft ', ' frm ', ' hrzn ', ' amp ']

def cleaning_text_basico(df, lista_elementos_eliminar):
    """
    Función generada para hacer una limpieza general del texto.
    
    param: df dataframe a limpiar
    :return: lista con el texto limpiado. Un elemento por cada fila del dataframe.
    """
    corpus = [' ' + text + ' ' for text in  df.text.tolist()] # Añadir espacio al principio y al final apra fcilitar limpieza
    corpus = [re.sub('@[A-Za-z0-9_]+',' ', text) for text in corpus] # Eliminar usuarios
    corpus = [re.sub(r'http\S+', '', text) for text in corpus] # Eliminas enlaces web
    corpus = [text.lower() for text in corpus] # Pasar a minúsculas
    #corpus = [clean_emojis(text) for text in corpus] # Eliminar emojis
    #corpus = [clean_emoticons(text) for text in corpus] # Eliminar emoticonos
    #corpus = [''.join(ch for ch in text if ch not in string.punctuation) for text in corpus] # Eliminar signos de puntuación
    corpus = [re.sub(r'[^a-z\s]', '', text) for text in corpus] # Eliminar lo que no sea minúscula o espacio
    for i in lista_elementos_eliminar:
        corpus = [re.sub(i,' ', text) for text in corpus] # Eliminar elementos de la lista
    corpus = [text.strip() for text in corpus] # Normalizar espacios
    corpus = [re.sub(r' +', ' ', text) for text in corpus] # Normalizar espacios
    corpus = [' '.join(word for word in text.split() if word not in stopwords.words('english')) for text in corpus] # Eliminar stopwords 
    return corpus

corpus_clean = cleaning_text_basico(df, lista_elementos_eliminar)

df['text_clean'] = corpus_clean

Se procede a realizar el análisis de sentimiento

vader_analyzer = SentimentIntensityAnalyzer()

def vader_sentiment_analysis(text):
    """
    Función para calcular el sentimiento de los tweets a partir de la librería VADER.
    
    param: text texto a calcular su sentimiento.
    :return: vader_sentiment resultado compound del análisis.
    """
    vader_sentiment_full = vader_analyzer.polarity_scores(text)
    vader_sentiment = vader_sentiment_full['compound']
    return vader_sentiment

df['vader_sentiment'] = df['text_clean'].apply(vader_sentiment_analysis)

def vader_sentiment_texto(score): 
    """
    Función para pasar a texto el resultado del análisis de sentimiento.
    
    param: score el resultado del análisis.
    :return: Valor en texto del resultado
    """
    if score < -0.05:
        return 'Negativo'
    elif score > 0.05:
        return 'Positivo'
    else:
        return 'Neutral'

df['vader_sentiment_analysis'] = df['vader_sentiment'].apply(vader_sentiment_texto)

Se obtienen los valores totales y porcentuales como se hizo anteriormente

data_total_vader = list()
data_porcentaje_vader = list()

for i in airlines :
    data1 = [list(df.loc[df.airline == i].vader_sentiment_analysis.value_counts().sort_index(ascending = False).values)+[df.loc[df.airline == i].shape[0]]]
    data_total_vader = data_total_vader+data1
    
    data2 = np.array(df.loc[df.airline == i].vader_sentiment_analysis.value_counts().sort_index(ascending = False).values)/df.loc[df.airline == i].shape[0]
    data2 = np.around(data2, decimals=4)*100
    data_porcentaje_vader.append(data2)  
    

data_total_vader = pd.DataFrame(data_total_vader, columns = columns1, index = airlines)

data_porcentaje_vader = pd.DataFrame(data_porcentaje_vader, columns = columns2, index = airlines)

df_total

data_total_vader

df_porcentaje

data_porcentaje_vader
